pyproject.toml
[tool.poetry]
name = "ai-ml-ops-challenge"
version = "0.1.0"
description = "AI & ML Ops Tech Challenge - Emporyum Tech Assistant"
authors = ["Candidate"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
langgraph = "^0.6.4"
langchain-core = "^0.3.0"
pydantic = "^2.0"
python-dotenv = "^1.0.0"
google-genai = "^1.65.0"
langchain-google-genai = "<3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
